Move LobbyPropertyReader to lobby project
