+++ lobby/src/main/java/games/strategy/engine/lobby/server/LobbyRunner.java
@@ -3,7 +3,7 @@ package games.strategy.engine.lobby.server;
 import java.util.logging.Level;
 
 import games.strategy.engine.config.FilePropertyReader;
-import games.strategy.engine.config.lobby.LobbyPropertyReader;
+import games.strategy.engine.lobby.server.config.LobbyPropertyReader;
 import games.strategy.util.ExitStatus;
 import lombok.extern.java.Log;
