+++ lobby/src/main/java/games/strategy/engine/lobby/server/ModeratorController.java
@@ -5,9 +5,9 @@ import java.util.Date;
 
 import javax.annotation.Nullable;
 
-import games.strategy.engine.config.lobby.LobbyPropertyReader;
 import games.strategy.engine.lobby.common.IModeratorController;
 import games.strategy.engine.lobby.common.IRemoteHostUtils;
+import games.strategy.engine.lobby.server.config.LobbyPropertyReader;
 import games.strategy.engine.lobby.server.db.BannedMacController;
 import games.strategy.engine.lobby.server.db.BannedUsernameController;
 import games.strategy.engine.lobby.server.db.Database;
